Continue transforming X86 parser/lexer
This is the inflexion point where I think it's necessary to split the
X86 parser into Intel and AT&T equivalents.  (There are just too many
overlaps between the two to produce a meaningful common parse tree.)

I'll probably focus on
one for now and come back to the other later on.