Add sanitisation to explainer
This change is fairly large, as it also comes with a new ability for
the sanitiser: various group of sanitisation functionality can be
switched on and off by passing a set of pass flags into the
sanitiser.

We currently just use this to enable a small set of semantically
stable sanitisations in the explainer, but this might get exposed on a
finer grain later on.