Move herd stanza to separate field in Config
The Litmusify command now respects this herd stanza, sort-of:
instead of -sendto /path/to/herd, now use -herd, and it'll pick up the
herd path from there.

If the stanza is missing, it'll default to 'herd7'.

Compiler stanzas still have a (herd) section, but it now takes a
Boolean and toggles enabling/disabling Herd processing.  The default
is now on.

This may break compiler.specs again---see the example for details.