First go at symbol mapping and Herd comparison
This bumper commit implements what appears from the outside to be a
successful scheme of mapping C-litmus locations through C symbols to
assembly symbols.  It's fairly rough---it needs better testing and
refactoring---but appears to get the job done.

This commit also changes the partial order types to contain the actual
items involved in superset/subset relations.