Separate abstractions from 'abstract kinds'.
Each abstraction now has a separate Kind module that enumerates
the high-level kinds of object that abstraction can be.  The usual
difference between `t` and `Kind.t` is that the latter has its
arguments removed.

This distinction makes it easier to do quick is-a tests on
abstract objects, makes it easier to add more detail to abstract
objects in future, and means that we can pretty-print abstract
elements and their types in different ways.