Replace Fold_map with traversable
Instead of implementing mapM in terms of fold_mapM, we now implement
fold_mapM in terms of mapM and the state monad.

This'll probably be a bit slower, but makes defining the traversable
much easier---we don't need to thread the accumulator through the
container anymore.

All traversable containers have been rearranged to suit.