Replace AST with a K&R-style C89 derivative
The parser here is almost 1:1 an implementation of 'The C Programming
Language', Second Edition.

This'll need some updating to get it up to C11 standard, but is
a closer facsimile than the Herd implementation.

Some things are missing in the lexer---float, string, char and
hex literals, as well as the lexer hack that'll be needed to
implement typedefs.