C: add state IDs to temp errors; fix litmus IDs
I haven't got around to writing proper error messages for the C parser
yet (there are over a hundred of them to do, and the parser'll probably
change a lot over the next month), but I've added each state number to
the temporary message so that I can look up errors slightly more easily
in the interim.

This also adds local-identifier support to Litmus postconditions, and
moves all of the Litmus AST into a separate module for eventual
functorisation.