Refactor to separate statement lists and programs
This mostly came to be by accident while trying to unify the
x86 and C litmus ASTs.  One thing led to another, and a large
amount of upheaval has been wrought.

This was quite a heavy set of changes, and is still quite rough
and ready in places.  Ah well.

Merry Christmas.