Io: strip out pipes and fds again
These aren't working, and trying to get them to work is too much of a
distraction at the moment, so I'll create an issue for them and use
temporary files for now.

This commit also implements basic automatic generation of temporary
files when passing stdin/stdout to a filter that doesn't support them,
but I haven't yet tested this to make sure it actually works.