Introduce validated C identifiers, etc
Most locations where C-style identifiers were expected now actively
validate for them.

Originally, this was supposed to be a quick generalisation of the
litmus identifier type in preparation for manipulating postconditions
without a full litmus AST in hand.  This turned into a massive rabbit
hole that took ages to step out of...