Trim down README
The number of `act` subcommands is getting quite high, and describing
them all in the readme is proving a bit unwieldy.  Instead, I'll
probably try describing them in depth on the wiki and in summary
in the tool itself for now, and use the README to point to both
locations.