Start adding mini-model path support
The idea here is that you construct a random path through a mini
program, and supply it to a fuzzer, which will then traverse the
path and do an operation at the end of it.

The idea is to try separate out as much random state as possible,
to make the fuzzers more easily testable.