Mini: Go deep into the rabbit hole
And go mad from the revelations within.

This was supposed to be a straightforward operation where I split
if-statements and statements, thus making it easier to implement
random generators on both by removing an anonymous record type,
and I'd quickly push it through and get back to actually building
a fuzzer.

Instead, it turned into a nightmare as OCaml's compiler repeatedly
shot down any attempt I made to push through the recursive
if-statement and statement modules, since various ppx-generated
bits of code were falling foul of its safe module restrictions.

After what felt like a lifetime of incoherent screaming and pleading
with the compiler, this is the compromise I've reached with it.

As a bonus, mini-model paths are now in a separate module and no
longer directly probe the structure of their target types, which
feels like a nice separation of concerns.

Can I make a fuzzer now, please?