Weighted_list: fix off-by-one error
And add tests to help make sure it doesn't happen again.

(The problem was that the sample range was 0..(max - 1), instead
of 0..max; this meant that the last item in the list effectively
had its weight reduced by 1, and, in the case of items whose
weight already _was_ 1, effectively removed them from the list).

A corollary of this is that the 'make-program' fuzzer action
now actually does something:

Fixes #72.