Finish fuzz action weight UX (hopefully)
The `act configure list-fuzzer-actions` command prints a pretty list
of available actions and their configured weights.

Added READMEs to actions, to make it easier for users to understand what
they do.

Haven't yet tested the weight modifiers, nor is there a programmatic way
to swarm the weight configuration, but that's all future work.