Asm: massive overhaul in prep for adding commands
While I'd hoped to have the stub generator in place today, this is no
mean refactor.

Moved all of the assembly jobs out of the Asm_job interface and
flipped them around so that they instead accept a job runner.  This
means that the job modules no longer have the responsibility of
enumerating all possible jobs, and makes it easier to add more.

I've also moved pretty the litmusifier and sanitiser to asm/.  This
leaves only the C preprocessor and sanitiser in lib; once I move the cpp
stuff to, say, act_common, then lib can be renamed to sanitiser/src.

This may have broken a lot of stuff.  Extreme caution!