Pare down set partial order
There's no reason why this should be included in all of the act set modules,
so I've refactored it to a more Base-like approach: most of it sits outside of
set modules now.

As the only actually-useful bit of My_set was the partial order stuff,
it's been renamed to `Set_partial_order`.  Tests have moved to the test
library.