Large refactoring and stub-gen commit of doom
Apparently I forgot to commit earlier, so this is yet another one of
those dreaded commits that changes a lot without explaining why.

Most of it is refactoring and autoformatting: for example, the new
'litmusifier_aux' stuff is me trying to pull tests out of the /src tree
and eliminate a functor.

I've also started adding more meat onto the stub generator; the outline
of an AST for the stubs has appeared, but not much is done with it yet.