Move filtering logic to new module and tidy up
The filtering stuff is becoming a right thorn in my side at this stage,
so I figured I'd begin cleaning it up.  Moving it into its own
'plumbing' module, with an eye to possibly eventually exporting it as an
opam package, seemed like a logical first step.  However, as these
things are wont to do, it apparently took way longer than I thought it
would.