(BROKEN) dump work on decoupling litmusifying
By which I mean: making sanitising/litmusifying its own separate filter,
and making it possible for act asm things to take in an asm/litmus test
directly.  The idea was to make it so that we could use the stub
generator as a direct driver for part of a simulator.

I've dumped this into a separate branch because honestly I'm not sure
how I'm going to finish it any time soon.  Everything is just an
absolute tangled mess, and I'd need a Herd7 x86 parser and everything
and oh god why did I start doing this.