Compiler: move from Config
I'm trying to hollow out the Config module so it only contains things
that actually pertain to the act.conf file and its care and feeding.  As
such, compilers, machines, and bits of stuff that depended too strongly
on them to be separated have gone to Compiler.

This needs some follow-up cleaning, since everything compiler-related is
in Compiler.Instance and needs breaking out into multiple modules.