Radically overhaul config (WIP)
Once again, another day's work has gone into a snowball of huge act
changes.  This pulls in a large amount of config changes, with the
salient points being:

- generalising the 'herd' and 'litmus' stanzas in preparation for more
  general simulator support;
- moving compiler and simulator config under machine config;
- adding groundwork for a more powerful defaults resolution system in
the future.

At the point of this commit, the new config system compiles, but doesn't
pass its tests, has a few quirks to iron out, and its compiler messages are outdated.

Sorry -- act is an nightmarish-to-maintain mess ATM, it seems.