(WIP) split off delitmus and formalise aux
This is another example of tugging at a thread and the whole garment
coming off, but such is the nature of act.

Start moving delitmusification to its own library, and make it output a
formal 'C litmus auxiliary record', which will eventually be passed into
the various assembly commands through a side-channel.

There's a lot of stubbing out in this commit, so the tests will compile
but won't work properly, and ACT itself will compile but won't work
properly.  I'll get back to this.