(WIP) replace C_variables with Var_map
The reason is that I'm trying to solidify the delitmusify aux format so
I can make it serialisable to JSON.  Once that's done, I can start
making tooling that operates on that JSON.

Some things still need doing, like wiring up the litmusifier so that it
can properly modify the aux file, and adding some more tests to Var_map.
However, this seems like a good place to rest for a bit.