Start script working on object splitmus
There's a bit of a blocker here: the postcondition emitted into the
skeleton refers to local variables by their fully qualified litmus ID,
but litmus thinks this means they're registers.

We should really be renaming them to t0r0 etc, but I've accidentally
architected the delitmusifier to not do this!