Port to OCaml 4.08
This includes a few bodge jobs here and there that I'll hopefully fix up
eventually.

Note that I've *disabled* OCaml 4.07 support, for now.  This is because
I don't trust myself to keep this working for OCaml 4.07 (indeed, I'll
probably start using the new module features immediately).  Homebrew
doesn't support OCaml 4.08 yet, so I'll follow this with a note in the
README.