Add initial fuzz checker and boilerplate
The fuzz checker only runs both sides of the fuzz pipeline through a
backend (eg herd) at the moment.  It should be straightforward to hook it up to
the diff-states tool, but presently I'm blocked by the fact that A) the
fuzzer generates Booleans, and B) herd doesn't support Booleans.  Natch.