State/diff: remove location mapping
The idea is that we'll eventually make new commands to do the location
mapping directly on observation JSON, on a case-by-case basis, for
example from an aux file.

The diff tool presently takes the intersection of the domains of both
sides of the state comparison.  This possibly hides errors where the
domains don't line up, and, as such, I want to add more reporting into
the diff tool as to which bits of each domain it's throwing away.

Gen_locmap still exists, but, if this new approach works well, is soon
to be deleted.