Migrate to Jane Street's yojson PPX
This is a slightly more robust PPX, and a slightly better fit with the
other Jane Street PPXen.  As  most of the PPXen we use are Jane Street
ones, it makes some sense to hop over.

It looks like `t_of_yojson'` will disappear in the next update,
though, so this'll need some cleanup later on.