Finish up initial sketch of do_memalloy_test
This still needs a little bit of TLC; for example, the fuzzer and tester
pipelines are interlinked in fuzz_and_test still, whereas they'd be
better off being split.  There could also be more control over whether
the fuzzer runs at all, and should definitely be a round of cleanup when
re-running the fuzzer to avoid cumulative fuzzer pile-ups and disk quota
oopses.

The slowness of sequential compiler running is starting to bite now,
too.  Maybe there's an easy-ish way to do this in parallel.