Strip out some redundant config predicates stuff
I haven't got round to it yet, but the `Act_config.Act` module is
now specifically configuration with machines resolved, and so
should be renamed at some point.  (It was always a weird name,
anyway).

The backend-specific bits of the toplevel should also take a
machine predicate to filter possible backends, but they presently
don't.