Finish command changes, add license command
Accidentally left most of the last commit hanging.  Oops.

Fixes #105.  (Inasmuch as there is now a specific command in most ACT
OCaml binaries that acknowledges the presence of herdtools7 code.  The
actual wording probably needs fixing.)

The README formatter has changed to one based on Fmt.  This is probably
a bit uglier with its word wrapping, but supports paragraphs properly.