Add metadata accounting to C_mini
C_mini metadata will eventually be used to instrument fuzz subjects with
dead-code, generated-or-existing, and other such flags, generalising the
(soon to be removed) `With_source.t` marker for top-level statements.

This is just the first bit of work on pushing metadata through; further
work will actually make it useful.

Some of the metadata integration is a bit janky, because of the way it
interacts with c-mini paths.  I might eventually monomorphise c-mini
paths and make them purely a fuzzer construct.