Move paths to fuzz
Step one in monomorphisation - in my opinion, paths are 'morally' a part
of c-mini, but they're only used in the fuzzer, and their
location-forced parametrisation over metadata is awkward.

As far as I know, nothing outside of the fuzzer will need path support.
If there ever comes a time where it does, then I'll have to change tack.