Add (unused) initial stabs at if-statement action
Also some path-based refactoring (more is sorely needed, the whole thing
is incredibly verbose and nasty to work with).

These type-check, and the various changes haven't obviously knackered
the tests, so all that remains to do is hook the sample 'duplicate
bodies' statement to the module list and watch it fall over repeatedly.