Move default weightings to config table
Also add the regression test from last commit because I'm terrible at
git!

The general idea here is that it's incredibly clumsy to have the
default weightings lodged in the module types and, for the price of a
little circumlocution in the code, we can fix that by moving the
default weights into the Config.ml main action table.