Fix 'N'
Specifically, teaches ACT that 'N' is not a valid Herd-safe identifier.
(Technically, it isn't a valid Litmus-safe identifier.  But teaching ACT
of the difference between Herd-safe and Litmus-safe identifiers would be
needlessly mindbogglingly complicated.)

More generally, this commit also teaches both C-identifier and Herd-identifier
validators about the concept of 'bad words' that cannot be valid
identifiers.  The list of bad words previously used by the C Quickcheck
generator now lives in the validator, though I had to remove `true` and
`false` as parts of ACT expect them to be valid identifiers.

Also adds a test to make sure 'N' is invalid.

Fixes #139.  (For now; it might re-open if other identifiers pose a
problem.)

Yes, this is a very over-engineered fix for the 'N' problem.