Overhaul compiler filtering
One might wonder, and honestly I've also wondered this: why overhaul the
compiler filtering?

I started doing this because I want to add a command for listing the
available *backends*, complete with filtering.  This will then be useful
for performing tasks like 'give me a local backend that behaves like
`herd`'.

In preparing to add said backend listing/filtering, I looked at the compiler
listing/filtering, and saw that it is a horrific and difficult to
generalise mess.  This bit of ACT has a lot of horrendous legacy cruft,
overly exuberant modularisation, and other nasty bits in it.  So, me
being me, I thought it wouldn't take too long to tidy it up a bit (in a
way that I can easily generalise to backends).

However, it _has_ taken quite a long time (1.5 days at least), and I
still _haven't_ even started the backend part of the equation, so... in
hindsight, this probably wasn't a good idea.

It's done now, though.