Add, and start using, `get_a_backend`
This script automagically bashes the ACT backend listing to find a
suitable backend of a given style.

I've started threading it through to replace 'herd' as the default in
pretty much everything that requires an arbitrary backend.  This both
fixes the issue of the make_conf script producing 'herd7' and there
being a mismatch, and, for bonus points, makes 'c_states' when a
compiler is given a lot more intuitive.

Rushing off to a meeting, so there may be more that needs doing.