Add return statements; refactor for expansion
The C-mini profile now understands expressionless returns.

In adding this, I figured that the number of 'primitive' statements, all
of which tend to be handled in very similar ways by large swathes of the
fuzzer at the moment, is now large enough to warrant factoring out
'primitive' expressions.  Much of this commit actually does that, rather
than just adding returns.