Instrument opt-gens with error information
Some opt-gen broke my last fuzzer run, so this is me scrabbling to try
to add enough information to start debugging it.

This commit also, for reasons of 'I was touching it anyway', teaches the
transform-list generator to be able to yield (n, 0) where n is the
length of the list.  This might not work properly yet, but the
generators are all broken anyway so I need to do some fixing.