Add missing loop support to paths & tidy up a bit
Hoo boy.  This is a massive commit, because it adds a heaping spoonful
of functionality that should always have been there, cleans up the path
module a little, renames a few things with misleading names, and... I
may have got minorly sidetracked adding pretty-printing to paths too,
but this is for use in tests later on.