Get stuff working enough to handcrank new method
Use PIC on non-assembly compiles, allow arbitrary flags to the compiler,
etc.

There needs to be extra work to tell litmus to produce a C harness,
since the current wrapper over litmus doesn't understand the idea of
providing *both* 'use C' and 'use x86', which it needs to do.

However, I've been able to get ACT to remotely compile a Litmus harness
(with some tweaking to ACT to teach it to be able to pass arbitrary
flags like 'pthread' in), copy the binary over, then copy it back and
xrun it to get an observation set.  This is so tantalisingly close to
what we need!

There are a few new utility modules.  I'm feeling like it'll soon be
worth splitting these off to Travesty and/or a new library.