Make aux files hold parameter position in param-mapped vars
Instead of a Boolean `mapped_to_global`, each var record now has a sum
type `mapped_to` that goes to either `["Global"]` or `["Param", n]`
where n is a hint as to where the variable would go in global
parameter maps.

This needs some work on the Python side to teach the Python libraries
about the new aux syntax.