Add more fine-grain delitmusifier control
The delitmusifier can now expressly be told what the name of the
implementation suffixes should be, as well as whether it should
qualify locals.  It defaults to not doing so for the former, and doing
so for the latter.

Also fixes an oddity with the delitmusifier in which, when qualifying,
it wouldn't qualify locals in param mode.  (Technically, doing so is
almost always pointless, but it feels orthogonal to do so.)