Make `make_conf` set thread cap
I've done this now because the Docker workflow depends very heavily on
being able to fire-and-forget `make_conf`, and so we can't manually
intervene to set a thread cap.

It's going to be interesting if we ever need to rewrite `make_conf` in
OCaml, because, as far as I'm aware, there's no analogue for
`multiprocessing.cpu_count()`.