Add dead-code store action
This was a bit of a journey - I changed my mind about how to factor out
the commonalities between store actions several times, and then, to get
the availability check to go through in a sensible manner, I had to add
a new path filter feature that does an availability check on a filter
over a subject.

The latter will come in handy in future actions as a way of making sure
the path filter and availability check are in step.

Needs a bit more testing, I feel - I've done a quick run of the fuzzer
to make sure that the fuzzer actually produces without erroring, but I
haven't actually checked it outputs sensible things.

Fixes #161.