Fix broken while loop cond traversal
This was manifesting as the delitmusifier failing to convert while loop
conditional lvalues to the appropriate type; turns out I made a boo-boo
when factoring out a base-map in statement and forgot to traverse over
these conditionals.  Oops.

I've added some small tests that'll check for regressions of this form.