Streamline C-mini environments
There is now no distinction between type and with-known-value
environments, and environments are generally values rather than modules
in an increasing amount of places.

The original idea for this was to make it easier to add things like
pulling out known values into generators.

Tests still pass but samples have permuted a bit.  I hope this is just
noise and not a sign of an issue.

Oh my God this took ages (two days, I kid you not!).