Make mem-order traversals roll back invalid values
This is a stopgap solution to an unsoundness issue whereby editing the
memory order of, say, a store with nested loads, would cause all of the
memory orders to be updated with the soundness checks appropriate to a
store.

This pushes the soundness check into the traversals, cleaning up a lot
of code and fixing the issue.  There is now a basic check in the test
for the above case.

Right now, mem-order traversals always occur recursively.  This might
not be the best idea (it limits atomicity), so at some point we might
want to revisit this.