Remove act-compiler
This is because act-tester has mostly subsumed the functionality of
act-compiler that isn't 'just run this compiler', and the latter isn't
useful enough to justify keeping two separate implementations of the
compiler code anymore.

act-machine, act-backend, and act-state will also be on the way out soon, but
they're presently part of the 'test amplification' hot path, which will
need some work to disentangle.

- Config files no longer accept compiler stanzas, compiler defaults, or
  arch defaults; just delete them.
- act-config probe has also been removed, because the above means it's not
  worth fixing up to deal with the lack of compilers.  Probing will
  appear on the act-tester version eventually.