Rename c-mini to FIR
FIR (fuzzable intermediate representation) is maybe a little more
telling as to what the intent of the language is.  It isn't necessarily
supposed to be a subset of C anymore, but instead a low common denominator
between potentially many C11-model programming languages arranged in a
way that makes fuzzing easier.