Generalise while loops to flow blocks
This took a while!

The idea is that a lot of the fuzzer logic that works for while loops
will also work for transactions (and eventually scoped locks and
for loops etc), and so instead of adding entirely new legs for each
we just capture a notion of a generalised 'block with added control
flow'.  If statements, being two blocks with added control flow,
remain as-is.

This'll need a bit of work on the reification and conversion front,
but this commit captures most of the legwork.