Spend lots of quality time with the LitmusC parser
The parser now supports atomics and synchronised blocks, and has a
couple of parsing error regressions to make sure those blocks have to be
blocks.

I've also added a few parser errors here and there, but the vast
majority of the errors remain undescribed (and probably will for some
time, as I'm really not sure how to get the parser into a shape that
doesn't make Menhir scream)

I've switched to using the 'open statement'/'closed statement' model for
eliminating the if-then-else shift/reduce.  This may or may not work
correctly, and may or may not need to be revised or wound back.

Lots of conflicts still remain, sadly.