Streamline outer interface of Path_producers
The only thing that was using path producers other than subject-level ones was
unit tests, which have been either converted to use subject-level tests
or binned (for now).

This simplification is an attempt to hide as much of the path-producers
interface as possible in preparation for significant refactoring of it.

I imagine a similar change will occur to Path_consumers soon, too.