Fix typo in transform-list generator
I'd accidentally made the generator skip past failing statements when
building ranges instead of aborting the generator; this meant that we
were generating transform-lists that straddled a bad statement, but not
ones that ended on one.

Fix was to replace unfold_step with unfold, changing the skip into a
None (ie, a 'done') in the same way that running off the end of a list produced
one.