Split fuzz into several libraries
Oh my days.  This was supposed to take at most an hour, and has instead
eaten the entire day.

The idea here was that there are now so many fuzzer actions that for
organisational reasons it makes sense to have them in a separate
library.  This also refocuses the 'fuzz' library as a sort-of virtual
machine for the actions to sit directly on top of, and the fuzzer
'runner' is separate for reasons of breaking a dependency cycle.

While I was here, I also tidied a few things up in the config, id, and
pp-by-reify libraries.