Add new 'payload gen' monad
This is a reader monad that replaces the state monad previously used for
generators.

Advantages at the moment include removing a lot of redundant parameter
passing, preventing generators from changing the fuzzer state, and
allowing for a cleaner API for fuzzer actions.

Eventually, I'd like to start unifying the generator and availability
stages, which tidying both up should make easier.