Move FIR reification/abstraction to litmus_c
This took much longer than expected... again.  Was not supposed to take
literally all day.

The idea here is that, eventually, we'll see litmus_c as just one
possible input/output dialect of FIR, and that FIR will be more of a
'lowest common denominator' between various languages that the ACT
system supports.

There is still a little too much coupling between FIR and litmus_c, but
I'll maybe clean it up later.