Bake variable checks into path filters
I think I must have accidentally made it so that storelikes could target
non-global variables, and the result is that the availability checks
that checked path existence and variable existence separately are no
longer adequate.

This is a quickfix that bungs in a new path filter constructor that
checks to make sure the path is constructing in a thread that has at
least one variable satisfying a set of predicates.

This needs cleaning up a little, but seems to work.