Replace parameter list equality with unification
Fixes #204.

There are a couple of issues here:

0) no tests yet;

1) there's duplication between the parameter unifying code in /litmus
and that in /fir, when there probably shouldn't be;

2) the new unifying code in /fir sorts the parameter lists, which might
break the delitmusifier.  So far, it passes all of its existing
regression tests, and I consider the delitmusifier mildly coderotted at
this stage, so I'll have to check when I next get back to using it.