Fix inability to take empty trace on right bisect
Right bisect usually has the semantics of 'the first trace that failed'
(whereas left bisect is 'the last trace that didn't fail').  It may well be
that every single trace considered by the bisection fails, in which case
the correct right bisect result should be the empty trace (as this is
the smallest valid trace).  Until this commit, the right bisect instead
only considered up to, and returned, the singleton trace.

This commit jiggles around the way the bisect works (removing some off
by ones and adding some elsewhere!) such that it considers the empty
trace in places it previously didn't.  It doesn't seem to affect any
results in our tests other than the right bisect of the empty trace, and
as such should be ok.

Fixes #205.