Overhaul op rule system
(THIS NEEDS MORE TESTING!)

Operator rules are now expressed as tables of the form (input, output),
where the input is either 'reflexive' (ie both inputs are the same) or
'left constant' or 'right constant' where constants are freeform.  The
output is either 'idempotent' (ie return the non-constant value) or
another constant.

This lets us theoretically express things such as:

(x * 0) -> 0
(x * 1) -> x
(x / x) -> 1
(x || true) -> true
.
.
.

The expression generator has been tweaked a bit so that wherever it was
previously looking for reflexive zeroes, it'll look for anything it can
make with the inputs that leads to zero, including things that involve
constants.

There is still a lot of stuff that could be done here, and a lot of
testing to do as well.