Fir: Expand assigns to include x++ and x--
Statements only; trying to keep track of expression *crements would both
be an utter nightmare from a undefined behaviour/sequence points
perspective *and* cause issues if we were to output Rust etc. later on.

The abstraction from Litmus/C also includes ++x and --x, under the
assumption that the two have compatible semantics as the *crements only
happen in statement position.

This is part of groundwork to introduce a lower-level for loop
representation.