Make cycle-breaker only break certain cycles
This is a pretty hacky solution to the problem of the new cycle breaker
in storelikes accidentally causing the cmpxchg generator to get stuck
(since we treat the object of the cmpxchg generator as both a dependency
generating source and a destination).

Basically, we now track whether we expect each expression to be 'safe'
or 'unsafe' in terms of whether it could possibly contain dependencies
that will break if used as the destination.  Cycle breaking only
considers unsafe expressions.

This'll need a bit more poking to make it lenient in the case of a
deadcode introduction.  BRB.