Add path flags for execute-multi
I'm still unconvinced that 'execute-multi' is actually the best name for
this, but I can't think of a better one.  (To be honest, the liveness
being 'Live' should probably change to 'Multi' too.  Or 'Many'.  Hmm.)

This also slightly changes the metadata API to bend it to the path flag
module's whims.