Slightly generalise and re-weight expr gen
This is in response to discussion with Ally and John whereby the
weighting of the expression generator towards producing large runs of
trivial constant expressions came up.  I've re-weighted some of the
decisions in favour of variable loads, added legs for boolean atomic
loading, and generally poked around on the expression generator a bit.

The main outstanding issue on the exprgen front is size; it's hard to
see where Base_quickcheck is making size decisions itself and where we
need to give it a push.