Add kv-never action
This is just a variant of kv-once that substitutes < and > for >= and <=
- quite nice in its simplicity.

Needs more testing, and the hurried way in which it was shoehorned in
means there are some rough edges to be sanded out - for instance, the
Insertion payload is now used for surrounds too, which suggests it's
time to generalise it and make it more widespread.

There are also probably a lot of things that are in modules and
shouldn't be, and vice versa.