Add anchors to path filters
An anchor in a path filter requires that that path 'sticks' to one
edge of its innermost block.  There are three anchors: anchor to top,
anchor to bottom, and 'full anchor' (anchor to top AND bottom - in
other words, span the entire block).  These, as well as the default
behaviour, form a lattice for the purposes of merging path filters.

I've shoehorned a fairly inelegant check onto both path producer and
consumer to make sure that any statement-list paths respect anchoring
if they point to the innermost nested block.

What are these useful for?  Nothing yet, but I intend to make a new
fuzzer action that inserts early-outs into loops anchored onto the end
of said loops; since those early-outs are always at the end of a loop,
they induce dead code.