Add early-out-loop-end fuzzer action
The idea here is that we jump out of a loop at the end of its current
end; doing so means that we statically know that nothing can be
executed below the early-out.

This needs a bit more work, ie making it auto-surround with an if-true
statement, to make it likely to produce interesting code.  However,
it's an interesting use of the anchoring.

(Needs testing, too.)