Add path filter contraption for direct block checks
This comes through the fact that the latest fuzzer bug is me conflating
the In_loop path flag (which literally means 'passes through a loop')
with 'ending in a loop'.  Rather than trying to get the path flag
system to tie path flags to specific bits of the path or anything like
that, I've just made a new type of path filter and tracking for it in
the path context.

This is... not very elegant.  It can probably be bashed into a nicer
form later on.