Move param-maps to state monad
This will pave the way for making the availability context available to
path filters in consumption position, allows possible use of the param
map outside of the payload generator, and I originally thought it'd be
useful to have the availability check outside of the randomised runner
but I'm not so sure now.

Also does a rather annoying amount of refactoring besides.

We'll need the param map to be passed into fuzzer traces before it's
safe to use it more widely.