Perform availability checks on choosing an action
The rough idea here is that, when we choose actions, we now loop over
the action table, picking up an action and testing its availability;
if it isn't available, we set its weight to zero in the table and
continue looping (erroring if we set every weight to zero).

Right now, once we pick an action, we set the availability back to
normal; we could do something more exciting later on.

Not only does this optimise for not needing to check the availability
of every single action - which is unusual when it happens -, but this
also pushes the availability check to somewhere where it's easier to
plug in other action sources (such as the mooted recommendation
queue).

Fixes #194.

We could do with changing the zeroing action so that it doesn't traverse
back over the list, really.