Move param-map back to runner monad
Yeah, this part-reverts d9b3077.  I was a bit silly and thought that
moving the param map into the state monad would make checking things on
the replay runner easier, or something like that; in fact, it means that
we'd have to lug the param map around in the trace as the meaning of the
trace would become dependent on it, and that seems like a bad idea.

Like any good revolution, while we've come full circle, we've moved on;
the path filters now just take fuzzer state instead of an availability
context, and that makes it much easier for us to check them at
consumption.  We've also still got the changes to the pool loop.