Add resetting to picker
This is a little ugly at the moment.  I intend to add a 'nop' action
and rearrange some of this so that the pool can report failure to
find an action, and the picker will compensate by filling in nop;
this will also help out when we do swarm testing as it'll protect us
from the issue of there being no available actions if the 'always
available' actions all get swarmed out.