Add 'accept recommendations' flag
This now means there are two rolls related to recommendations: one
to decide whether the recommendation actually hits the queue, and
another to decide whether to take it at a particular moment in time.

This helps both break the recommendations up across time, but also
make sure that we don't inundate the queue with lots of recommendations
from one action.