Add rudimentary swarm testing
The initial action deck now gets pruned to a particular amount of
actions (currently 20), which are then the only actions that can be
used both in recommendations and in normal operation.

Recommendations now no longer error if they don't exist in the deck,
as a consequence.  This might mask errors in the future, but hopefully
not if we program them defensively.

I had to implement my own sampling logic because the Base stuff uses
non-splittable randoms, which is painful - but there are some
rudimentary tests.  They should really have proper quickchecks but
effort.