Push anchor metadata through path p/c
Path contexts now contain the most recent computed anchor as metadata,
and checks on anchor safety now occur at generation site.  The
path generation samples show that the anchor passes neatly through into
the metadata attached to the generated paths.

This makes it easier to do #220; we still currently generate then
reject paths that make no sense for anchor blocks, but it should be
much easier to add the special-casing now (if we can exfiltrate the
anchor from the path filter).