Allow break if the loop is empty
We encode this as checking if the insertion point has top anchoring;
as we already require it to have bottom anchoring, this is tantamount
to checking for full anchoring, which *itself* is tantamount to checking
that the loop is empty.

Amended the Morisset example so as to use break rather than continue,
thereby showing that we can at least validate that bug stimulus
(whether we can generate it is somewhat more difficult to prove).