Limit things to make unit tests reasonable
Lots of size divisions, and merging tests.

This probably massively thwarts the tendency of the expression
generator to generate intricately nested expressions, but hopefully the
increasing horizontal expressivity'll help?  Either way, we need to
profile and optimise the expression generator - it is extremely slow
and the recursion I've added lately has made it much, much worse.