Add variable legs to constant generator
These cook up a load, possibly transform it to massage the known value
into the desired constant value, and substitute it for a constant.
For instance, if we know x has known value 42, we can use (x - 43) to
generate -1.

This overlaps a little with the negation leg and the (x - x = 0) rule,
but c'est la vie.